Story�Number:�	OAS-213
Description Of Story:�	All type of users will be able to view the response and result of individual student in view status page.

Hints To QA For Testing:�	Login to TAS for TASC Readiness user and check for response and score details in 'Score Details' accordion of View Status page for individual rosters.

�Developer Comments:� The TASC Readiness customer should have the "View_Response_Result" configuration and default value is 'T'. This feature should be checked in various TAS UI supported browsers.